Issue-178: Change feedback bar color
