+++ src/Feedback.tsx
@@ -1,12 +1,11 @@
 import {ShowChartOutlined} from "@mui/icons-material";
 import InfoOutlined from "@mui/icons-material/InfoOutlined";
-import {Container, DialogContentText, Grid, LinearProgress, Paper, useTheme} from "@mui/material";
-import {blue, green, red, yellow} from "@mui/material/colors";
-import {ChartOptions} from "chart.js";
+import {Container, DialogContentText, Grid, LinearProgress, Paper} from "@mui/material";
+
 import 'chart.js/auto';
 import ChartDataLabels from 'chartjs-plugin-datalabels';
 import {isPast, isValid} from "date-fns";
-import React, {useCallback, useEffect, useMemo, useState} from 'react';
+import React, {useCallback, useEffect, useState} from 'react';
 import {Chart} from "react-chartjs-2";
 import {useTranslation} from "react-i18next";
 import {Redirect, useLocation} from "react-router-dom";
@@ -18,112 +17,8 @@ import ResponsiveIconButton from "./common/ResponsiveIconButton";
 import RetryMessage from "./common/RetryMessage";
 import useDefaultTracking from "./common/Tracking";
 import {useSnackBar} from "./common/UseSnackBar";
+import {useBarChartData, useBarChartOptions, usePieChartData, usePieChartOptions} from "./feedback/Charts";
 import {FeedbackModel} from "./service/Model";
-
-const useBarChartData = (data: { self: number, others: number }) => {
-    const theme = useTheme();
-    const {t} = useTranslation();
-    return useMemo(() => ({
-        labels: [t('feedback_energy_consumption')],
-        datasets: [
-            {
-                label: t('feedback_my_energy_consumption'),
-                data: [data.self],
-                backgroundColor: red["800"],
-                borderColor: theme.palette.background.paper,
-            },
-            {
-                label: t('feedback_other_energy_consumption'),
-                data: [data.others],
-                backgroundColor: blue["800"],
-                borderColor: theme.palette.background.paper,
-            },
-        ],
-    }), [theme, data.self, data.others, t]);
-};
-const usePieChartData = (data: { high: number, med: number, low: number }) => {
-    const theme = useTheme();
-    const {t} = useTranslation();
-    return useMemo(() => ({
-        labels: [t('feedback_high_energy'), t('feedback_med_energy'), t('feedback_low_energy')],
-        datasets: [
-            {
-                data: [data.high, data.med, data.low].map(v => v * 100),
-                backgroundColor: [green["800"], yellow["800"], red["800"]],
-                borderColor: theme.palette.background.paper,
-            },
-        ],
-    }), [theme, data.high, data.med, data.low, t]);
-}
-const useBarChartOptions: () => ChartOptions<"bar"> = () => {
-    const theme = useTheme();
-    return useMemo(() => ({
-        responsive: true,
-        scales: {
-            xAxis: {display: false},
-            yAxis: {
-                position: "right",
-                grid: {
-                    drawBorder: false,
-                    color: theme.palette.divider
-                },
-                ticks: {
-                    callback: tickValue => tickValue + " kWh",
-                    padding: 10,
-                    stepSize: 1,
-                    color: theme.palette.text.primary,
-                    font: {family: theme.typography.fontFamily}
-                }
-            },
-        },
-        plugins: {
-            legend: {
-                position: 'bottom',
-                labels: {
-                    font: {family: theme.typography.fontFamily},
-                    color: theme.palette.text.primary
-                }
-            },
-            tooltip: {enabled: false},
-            datalabels: {
-                color: theme.palette.primary.contrastText,
-                font: {
-                    family: theme.typography.fontFamily,
-                    size: 20,
-                },
-                formatter: value => value.toFixed(1) + " kWh"
-            }
-        }
-    }), [theme]);
-}
-const usePieChartOptions: () => ChartOptions<"pie"> = () => {
-    const theme = useTheme();
-    return useMemo(() => ({
-        responsive: true,
-        scales: {
-            xAxis: {display: false},
-            yAxis: {display: false}
-        },
-        plugins: {
-            legend: {
-                position: 'bottom',
-                labels: {
-                    font: {family: theme.typography.fontFamily},
-                    color: theme.palette.text.primary
-                }
-            },
-            tooltip: {enabled: false},
-            datalabels: {
-                color: theme.palette.primary.contrastText,
-                font: {
-                    family: theme.typography.fontFamily,
-                    size: 20,
-                },
-                formatter: value => value.toFixed(1) + "%"
-            }
-        }
-    }), [theme]);
-}
 
 function Feedback(props: PrivateRouteProps) {
     const {Track} = useDefaultTracking({page: 'Feedback'});
